requirements.txt
fastapi
uvicorn[standard]
face_recognition
python-multipart
pillow
numpy